Consolidate and optimize UI tests for Swift example app
- Reduced test count from 22 to 10 by grouping related operations
- Organized tests by wallet type (EVM, Solana, Cosmos) to share account creation
- Added proper test coverage for Cosmos proto/amino signing methods
- Added transaction signing test for Solana wallets
- Fixed test assertions to properly validate success (tests now fail on errors)
- Removed duplicate message signing tests
- Implemented shared logout helper to ensure clean state between tests
- Updated test plan configuration for parallel execution
- Expected runtime reduction from ~22 minutes to ~5-7 minutes

Each wallet type test class now creates an account once and shares it across
all tests in that class, significantly reducing redundant setup time.
